Improve game rendenring and client/server communication
There are still errors, but most of the previous bugs have been fixed. The client doesn't receive chunk packets from server, must have a look at "PlayerChunkMap".

+++ openverse/src/main/java/xyz/upperlevel/openverse/network/world/PlayerChangeLookPacket.java
@@ -13,17 +13,17 @@ import xyz.upperlevel.hermes.Packet;
 @NoArgsConstructor
 @AllArgsConstructor
 public class PlayerChangeLookPacket implements Packet {
-    private float yaw, pitch;
+    private double yaw, pitch;
 
     @Override
     public void toData(ByteBuf out) {
-        out.writeFloat(yaw);
-        out.writeFloat(pitch);
+        out.writeDouble(yaw);
+        out.writeDouble(pitch);
     }
 
     @Override
     public void fromData(ByteBuf in) {
-        yaw = in.readFloat();
-        pitch = in.readFloat();
+        yaw = in.readDouble();
+        pitch = in.readDouble();
     }
 }
